KDtree_distance_threshold_mum int to float

+++ chromapylot/parameters/matrix_params.py
@@ -17,7 +17,7 @@ class MatrixParams:
         masks2process (Dict[str, str]): Dictionary mapping mask names to their corresponding channels. Default is {'nuclei': 'DAPI', 'mask1': 'mask0'}.
         flux_min (int): Minimum flux required to keep an object. Default is 10.
         flux_min_3D (float): Minimum flux required to keep an object in 3D. Default is 0.1.
-        KDtree_distance_threshold_mum (int): Distance threshold used to build KDtree. Default is 1.
+        KDtree_distance_threshold_mum (float): Distance threshold used to build KDtree. Default is 1.
         toleranceDrift (Union[int, List[int]]): ZXY tolerance used for block drift correction, in pixels. Default is [3, 1, 1].
         remove_uncorrected_localizations (bool): Flag indicating whether to remove uncorrected localizations. Default is True.
         z_offset (float): Z offset value. Default is 2.0.
@@ -33,7 +33,7 @@ class MatrixParams:
     )
     flux_min: int = set_default("flux_min", 10)
     flux_min_3D: float = set_default("flux_min_3D", 0.1)
-    KDtree_distance_threshold_mum: int = set_default("KDtree_distance_threshold_mum", 1)
+    KDtree_distance_threshold_mum: float = set_default("KDtree_distance_threshold_mum", 1)
     toleranceDrift: Union[int, List[int]] = set_default("toleranceDrift", [3, 1, 1])
     remove_uncorrected_localizations: bool = set_default(
         "remove_uncorrected_localizations", True
